Unnecessary unboxing

+++ src/main/java/net/sf/antcontrib/cpptasks/types/FlexLong.java
@@ -55,7 +55,7 @@ public class FlexLong {
      * @return long
      */
     public long longValue() {
-        return value.longValue();
+        return value;
     }
 
     /**
